Attempt to address issue #9, by syncronising the loops in each thread. In the Main class, I create a CyclicBarrier, initialised with the parameter 4, then in GameThread, at the end of each iteration, I call barrier.await(). This stops each thread until 4 'await()'s have been called, then they all start again.
I don't know if this is actually necessary, and it might be a massive performance disaster, but it seems to make sense to me. Also, I've no idea how (if?) this clashes with the LWJGL Display.sync() call in the renderer's loop, which must do something similar, but with time.