Issue 28 Well, this might be better or worse or the same. Probably better though. The entities list in assContainer was only used by the collision detection, and I'm not sure that there are likely to be many other scenarios where we'd want a list of all entities of all types, so I've made the CollisionThread responsible for building its own collection of assets that it will test for collision.
To get around the null models, I've added a 'readyForCollisionDetection()' method, which just tests if the model is null, but presumably we might want other setup stuff to be considered in there eventually.

This does mean we're still re-building the list of entities each time round the collision detection loop, but if it becomes a problem, we can maybe do some caching or something.